Added a reisterUser method which currently assumes that all registering users are tenants.
Updated signup page to register this user. So it goes into two collections, one of users which is managed by meteor itself and then based on role it is put into the Tenant collection created by us (other fields are left blank which can be filled in later under profile settings of the user)

the password field is removed from Tenants (same to be done with others) as this is handled by meteor and storing password can cause security issues.